+++ reference/versions/index.js
@@ -12,8 +12,9 @@ var versions = {
     "0.16.0": require("./0.16.0/reference.json"),
     "0.17.0": require("./0.17.0/reference.json"),
     "0.18.0": require("./0.18.0/reference.json"),
-    "0.19.0": require("./0.19.0/reference.json")
+    "0.19.0": require("./0.19.0/reference.json"),
+    "0.20.0": require("./0.20.0/reference.json")
 };
 
-versions.latest = versions["0.19.0"];
+versions.latest = versions["0.20.0"];
 module.exports = versions;